Split out resource classes
